invent the ability to equip and unequip items

+++ Player.pm
@@ -56,12 +56,12 @@ sub inventory {
     foreach my $item ( @inventory ) {
         # my $how_many = $possessions->{$item} == 1 ? 'a' : $possessions->{$item};
         # $item .= 's' if $how_many ne 'a';
-        print "\tYou have a $item\n";
+        print "\tYou have a $item in your pack\n";
     }
     foreach my $item ( @equipment ) {
         # my $how_many = $possessions->{$item} == 1 ? 'a' : $possessions->{$item};
         # $item .= 's' if $how_many ne 'a';
-        print "\tYou have a $item equipped\n";
+        print "\tYou have a $item in your hand\n";
     }
 }
 
@@ -197,6 +197,7 @@ sub _kill {
     return warn "\t\u${word} who with what?\n" unless ref $baddie;
     return warn "\t\uWhat will you kill the $baddie with?\n" unless ref $item;
     return warn "\tYou don't have a $item\n" unless $self->has($item);
+    return warn "\tYou must equip your $item before you may use it\n" unless $self->has_on($item);
     return warn "\tThe $baddie is not something that can be killed\n" unless $baddie->get_health();
     return warn "\tThere is no $baddie here\n" unless $baddie->where() eq $here;
     return warn "\tWhy would you kill the poor innocent $baddie?\n"
